+++ ChangeLog.md
@@ -2,7 +2,7 @@
 
 ## Unreleased changes
 
-- Extend the `GET /v0/transactionCost` endpoint to support `type=simplePltTransfer`, which takes `tokenId`, `tokenAmount`, and `textMemo` as query parameters to compute the cost of a simple PLT transfer.
+- Extend the `GET /v0/transactionCost` endpoint to support `type=tokenUpdate`, which takes `tokenId`, `listOperationsSize`, and `tokenOperationTypeCount` as query parameters to compute the cost of token update transactions.
 
 ## [0.39.0] - 2025-07-15